LSTM training
I am treating each day as a batch, so the batch size changes every iteration. Since LSTM state is reset every iteration anyways, I'm rebuilding it each iteration with the current batch size. The prediction for the next day is just the average of the onehot vectors, with a configurable window.